# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Karayolu Evrak Akıllı Ajan REST API",
    "description": "Bağımsız frontend istemcileri için kaynak doğrulamalı evrak işleme servisi",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8010",
      "description": "Yerel geliştirme backend'i"
    }
  ],
  "paths": {
...
    "/api/v1/processes/text": {
      "post": {
        "tags": [
          "processes"
        ],
        "summary": "Metin evrakı işle",
        "operationId": "processTextDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/file": {
      "post": {
        "tags": [
          "processes"
        ],
        "summary": "Dosya evrakı yükle ve işle",
        "operationId": "uploadDocument",
        "parameters": [
          {
            "name": "compile_pdf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Compile Pdf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadDocument"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/{document_id}": {
      "get": {
        "tags": [
          "processes"
        ],
        "summary": "Evrak sürecini getir",
        "operationId": "getProcess",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/{document_id}/information": {
      "post": {
        "tags": [
          "processes"
        ],
        "summary": "Eksik evrak bilgilerini tamamla",
        "operationId": "updateProcessInformation",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InformationUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/{document_id}/approval": {
      "post": {
        "tags": [
          "processes"
        ],
        "summary": "Evrak taslağını onayla",
        "operationId": "approveProcess",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/{document_id}/artifacts/tex": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "LaTeX artifact'ını indir",
        "operationId": "downloadLatexArtifact",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/{document_id}/artifacts/pdf": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "PDF artifact'ını indir",
        "operationId": "downloadPdfArtifact",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApprovalRequest": {
        "properties": {
          "approved_by": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Approved By"
          }
        },
        "type": "object",
        "required": [
          "approved_by"
        ],
        "title": "ApprovalRequest"
      },
...
      "Body_uploadDocument": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_uploadDocument"
      },
...
      "DocumentLayout": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/DocumentLine"
            },
            "type": "array",
            "title": "Lines"
          },
          "page_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Count",
            "default": 1
          },
          "coordinate_system": {
            "type": "string",
            "enum": [
              "normalized_page",
              "unavailable"
            ],
            "title": "Coordinate System",
            "default": "unavailable"
          }
        },
        "type": "object",
        "title": "DocumentLayout",
        "description": "Coordinate-aware document text. LLMs may cite IDs but never set bboxes."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InformationUpdateRequest": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "minProperties": 1,
            "title": "Fields"
          },
          "compile_pdf": {
            "type": "boolean",
            "title": "Compile Pdf",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "InformationUpdateRequest"
      },
...
      "ProcessEvent": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ProcessStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "agent"
        ],
        "title": "ProcessEvent"
      },
      "ProcessState": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "status": {
            "$ref": "#/components/schemas/ProcessStatus",
            "default": "alindi"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage",
            "default": "evrak_alimi"
          },
          "completed_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Completed Steps"
          },
          "pending_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending Actions"
          },
          "missing_information": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Information"
          },
          "next_step": {
            "type": "string",
            "title": "Next Step",
            "default": "Evrakın işlenmesini başlatınız."
          },
          "possible_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Possible Actions"
          },
          "provided_information": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Provided Information"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "document_layout": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentLayout"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "layer1_audit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Layer1Audit"
              },
              {
                "type": "null"
              }
            ]
          },
          "search_hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Search Hits"
          },
          "retrieval_diagnostics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalDiagnostics"
              },
              {
                "type": "null"
              }
            ]
          },
          "graph_decision_trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphDecisionTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "llm_trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMRunTrace"
              },
              {
                "type": "null"
              }
            ]
          },
          "verified_references": {
            "items": {
              "$ref": "#/components/schemas/VerifiedReference"
            },
            "type": "array",
            "title": "Verified References"
          },
          "template_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemplateDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingRecommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DraftPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "compliance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComplianceResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ProcessEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "ProcessState"
      },
      "ProcessStatus": {
        "type": "string",
        "enum": [
          "alindi",
          "okunuyor",
          "siniflandiriliyor",
          "mevzuat_araniyor",
          "kaynak_dogrulaniyor",
          "eksik_bilgi_bekleniyor",
          "yazi_turu_seciliyor",
          "birim_yonlendiriliyor",
          "taslak_hazirlaniyor",
          "uygunluk_kontrolunde",
          "kullanici_onayi_bekleniyor",
          "tamamlandi",
          "hata"
        ],
        "title": "ProcessStatus"
      },
...
      "TemplateDecision": {
        "properties": {
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "user_approval_required": {
            "type": "boolean",
            "title": "User Approval Required",
            "default": true
          },
          "alternatives": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Alternatives"
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "template_id",
          "rationale",
          "confidence"
        ],
        "title": "TemplateDecision"
      },
      "TextProcessRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Text"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name",
            "default": "kullanici_metni.txt"
          },
          "compile_pdf": {
            "type": "boolean",
            "title": "Compile Pdf",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextProcessRequest"
      },
...
    }
  }
}